Revert Stage counter to unsigned short
Revert the unsigned int widening — the overflow was a theoretical
concern, not an observed issue.

+++ modern.src/stage.h
@@ -44,7 +44,7 @@ class StageClass {
 		**	This handles the animation stage of the object. This includes smoke, walking,
 		**	flapping, and rocket flames.
 		*/
-		unsigned int Stage;
+		unsigned short Stage;
 
 		/*
 		**	This is the countdown timer for stage animation. When this counts down
